Announcing general availability of Ai1EC v2.1.6.
git-svn-id: https://plugins.svn.wordpress.org/all-in-one-event-calendar/tags/2.1.6@1028999 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/theme/compiler.php
@@ -147,7 +147,10 @@ class Ai1ec_Theme_Compiler extends Ai1ec_Base {
 		if ( ! $this->_prune_dir( $parent ) ) {
 			return false;
 		}
-		if ( mkdir( $cache_dir, 0754, true ) ) {
+		if (
+			is_dir( $cache_dir ) && chmod( $cache_dir, 0754 )
+			|| mkdir( $cache_dir, 0754, true )
+		) {
 			return true;
 		}
 		return false;
@@ -198,6 +201,9 @@ class Ai1ec_Theme_Compiler extends Ai1ec_Base {
 			}
 		}
 		closedir( $handle );
+		if ( is_file( $cache_dir . DIRECTORY_SEPARATOR . 'EMPTY' ) ) {
+			return true; // ignore, this directory is intentionally here
+		}
 		return rmdir( $cache_dir );
 	}
